(refactoring) Get rid of template change sender
