+++ ydb/core/change_exchange/change_exchange.h
@@ -7,50 +7,6 @@
 #include <ydb/core/scheme/scheme_pathid.h>
 
 #include <util/generic/vector.h>
-#include <util/string/builder.h>
-#include <util/string/join.h>
-
-#include <memory>
-#include <variant>
-
-namespace NKikimr {
-
-namespace NDataShard {
-    class TChangeRecord;
-}
-
-namespace NReplication::NService {
-    class TChangeRecord;
-}
-
-namespace NBackup::NImpl {
-    class TChangeRecord;
-}
-
-struct IChangeRecordContainer {
-    virtual ~IChangeRecordContainer() = default;
-    virtual TString Out() const = 0;
-};
-
-template <typename T>
-struct TBaseChangeRecordContainer: public IChangeRecordContainer {
-    TVector<typename T::TPtr> Records;
-
-    TBaseChangeRecordContainer() = default;
-
-    explicit TBaseChangeRecordContainer(TVector<typename T::TPtr>&& records)
-        : Records(std::move(records))
-    {}
-
-    TString Out() const override {
-        return TStringBuilder() << "[" << JoinSeq(",", Records) << "]";
-    }
-};
-
-template <typename T>
-struct TChangeRecordContainer {};
-
-}
 
 namespace NKikimr::NChangeExchange {
 
@@ -117,26 +73,11 @@ struct TEvChangeExchange {
     };
 
     struct TEvRecords: public TEventLocal<TEvRecords, EvRecords> {
-        using TChangeRecordVector = std::variant<
-            std::shared_ptr<TChangeRecordContainer<NDataShard::TChangeRecord>>,
-            std::shared_ptr<TChangeRecordContainer<NReplication::NService::TChangeRecord>>,
-            std::shared_ptr<TChangeRecordContainer<NBackup::NImpl::TChangeRecord>>
-        >;
+        TVector<IChangeRecord::TPtr> Records;
 
-        TChangeRecordVector Records;
-
-        explicit TEvRecords(TChangeRecordVector&& records);
+        explicit TEvRecords(const TVector<IChangeRecord::TPtr>& records);
+        explicit TEvRecords(TVector<IChangeRecord::TPtr>&& records);
         TString ToString() const override;
-
-        template <typename T>
-        static TEvRecords* New(TVector<typename T::TPtr>&& records) {
-            return new TEvRecords(std::make_shared<TChangeRecordContainer<T>>(std::move(records)));
-        }
-
-        template <typename T>
-        inline auto& GetRecords() {
-            return std::get<std::shared_ptr<TChangeRecordContainer<T>>>(Records)->Records;
-        }
     };
 
     struct TEvForgetRecords: public TEventLocal<TEvForgetRecords, EvForgetRecods> {
