+++ ydb/core/change_exchange/fwd.h
@@ -1,0 +1,17 @@
+#pragma once
+
+namespace NKikimr {
+
+namespace NDataShard {
+    class TChangeRecord;
+}
+
+namespace NReplication::NService {
+    class TChangeRecord;
+}
+
+namespace NBackup::NImpl {
+    class TChangeRecord;
+}
+
+}
