+++ ydb/core/change_exchange/ya.make
@@ -3,8 +3,9 @@ LIBRARY()
 SRCS(
     change_exchange.cpp
     change_record.cpp
+    change_sender.cpp
     change_sender_monitoring.cpp
-    change_sender_partitioner.cpp
+    resolve_partition.cpp
 )
 
 GENERATE_ENUM_SERIALIZATION(change_record.h)
@@ -13,7 +14,6 @@ PEERDIR(
     ydb/core/base
     ydb/core/scheme
     ydb/library/actors/core
-    ydb/library/yverify_stream
     library/cpp/monlib/service/pages
 )
