+++ ydb/core/tx/datashard/change_record.h
@@ -2,15 +2,12 @@
 
 #include "datashard_user_table.h"
 
-#include <ydb/core/change_exchange/change_record.h>
-#include <ydb/core/change_exchange/change_sender_resolver.h>
 #include <ydb/core/change_exchange/change_exchange.h>
+#include <ydb/core/change_exchange/change_record.h>
 #include <ydb/core/scheme/scheme_pathid.h>
 #include <ydb/core/scheme/scheme_tablecell.h>
-#include <ydb/services/lib/sharding/sharding.h>
 
 #include <util/generic/maybe.h>
-#include <util/string/join.h>
 
 namespace NKikimrChangeExchange {
     class TChangeRecord;
@@ -24,8 +21,6 @@ class TChangeRecord: public NChangeExchange::TChangeRecordBase {
     friend class TChangeRecordBuilder;
 
 public:
-    using TPtr = TIntrusivePtr<TChangeRecord>;
-
     ui64 GetGroup() const override { return Group; }
     ui64 GetStep() const override { return Step; }
     ui64 GetTxId() const override { return TxId; }
@@ -45,6 +40,8 @@ public:
     i64 GetSeqNo() const;
     TInstant GetApproximateCreationDateTime() const;
     bool IsBroadcast() const override;
+
+    void Accept(NChangeExchange::IVisitor& visitor) const override;
 
     void Out(IOutputStream& out) const override;
 
@@ -123,23 +120,10 @@ public:
 
 }; // TChangeRecordBuilder
 
-}
-
-namespace NKikimr {
-
-template <>
-struct TChangeRecordContainer<NDataShard::TChangeRecord>
-    : public TBaseChangeRecordContainer<NDataShard::TChangeRecord>
-{
-    using TBaseChangeRecordContainer<NDataShard::TChangeRecord>::TBaseChangeRecordContainer;
-};
+NChangeExchange::IPartitionResolverVisitor* CreateDefaultPartitionResolver(const NKikimr::TKeyDesc& keyDesc);
 
 }
 
 Y_DECLARE_OUT_SPEC(inline, NKikimr::NDataShard::TChangeRecord, out, value) {
     return value.Out(out);
-}
-
-Y_DECLARE_OUT_SPEC(inline, NKikimr::NDataShard::TChangeRecord::TPtr, out, value) {
-    return value->Out(out);
 }
