+++ ydb/core/tx/datashard/datashard_ut_change_exchange.cpp
@@ -2,7 +2,7 @@
 #include "datashard_ut_common_kqp.h"
 
 #include <ydb/core/base/path.h>
-#include <ydb/core/change_exchange/change_sender_common_ops.h>
+#include <ydb/core/change_exchange/change_sender.h>
 #include <ydb/core/persqueue/events/global.h>
 #include <ydb/core/persqueue/user_info.h>
 #include <ydb/core/persqueue/write_meta.h>
