+++ ydb/core/tx/replication/service/json_change_record.h
@@ -4,7 +4,6 @@
 
 #include <ydb/core/change_exchange/change_exchange.h>
 #include <ydb/core/change_exchange/change_record.h>
-#include <ydb/core/change_exchange/change_sender_resolver.h>
 #include <ydb/core/scheme/scheme_tablecell.h>
 #include <ydb/core/scheme_types/scheme_type_info.h>
 #include <ydb/core/tablet_flat/flat_row_eggs.h>
@@ -29,7 +28,6 @@ class TChangeRecord: public NChangeExchange::TChangeRecordBase {
     friend class TChangeRecordBuilder;
 
 public:
-    using TPtr = TIntrusivePtr<TChangeRecord>;
     using TBuilder = TChangeRecordBuilder;
 
     struct TSerializationContext {
@@ -44,21 +42,19 @@ public:
         {}
     };
 
-    const static NKikimrSchemeOp::ECdcStreamFormat StreamType = NKikimrSchemeOp::ECdcStreamFormatJson;
-
     ui64 GetGroup() const override;
     ui64 GetStep() const override;
     ui64 GetTxId() const override;
     EKind GetKind() const override;
-    TString GetSourceId() const;
 
     void Serialize(NKikimrTxDataShard::TEvApplyReplicationChanges_TChange& record, TSerializationContext& ctx) const;
 
     TConstArrayRef<TCell> GetKey(TMemoryPool& pool) const;
     TConstArrayRef<TCell> GetKey() const;
 
+    void Accept(NChangeExchange::IVisitor& visitor) const override;
+
 private:
-    TString SourceId;
     NJson::TJsonValue JsonBody;
     TLightweightSchema::TCPtr Schema;
 
@@ -91,21 +87,6 @@ public:
 
 }
 
-namespace NKikimr {
-
-template <>
-struct TChangeRecordContainer<NReplication::NService::TChangeRecord>
-    : public TBaseChangeRecordContainer<NReplication::NService::TChangeRecord>
-{
-    using TBaseChangeRecordContainer<NReplication::NService::TChangeRecord>::TBaseChangeRecordContainer;
-};
-
-}
-
 Y_DECLARE_OUT_SPEC(inline, NKikimr::NReplication::NService::TChangeRecord, out, value) {
     return value.Out(out);
-}
-
-Y_DECLARE_OUT_SPEC(inline, NKikimr::NReplication::NService::TChangeRecord::TPtr, out, value) {
-    return value->Out(out);
 }
